speech_to_text_transcriber/requirements.txt
google-cloud-speech==2.7.0
ffmpeg-python==0.2.0
Flask==2.0.1
requests==2.26.0
gunicorn==20.1.0